drivers: qcom: subsys_backup: Fix variable type
On 64 bit arch, size_t (aka unsigned long *) becomes 64 bits.
Use u32 variable type instead to fix compile time errors on 64 bit
architectures.

+++ drivers/soc/qcom/subsys_backup.c
@@ -1291,7 +1291,7 @@ static int subsys_backup_driver_probe(struct platform_device *pdev)
 {
 	int ret;
 	struct subsys_backup *backup_dev;
-	size_t buf_size;
+	u32 buf_size;
 
 	backup_dev = devm_kzalloc(&pdev->dev, sizeof(struct subsys_backup),
 					GFP_KERNEL);
